Enable S3 caching, see comment below
This cache will only hit when he browser literally reuses a link. S3 generates a
new link every time and for browser caches to hit links must be completely
identical. If we want more caching, we need to play games with the timestamps in
the link. Essentially pretend that the link was generated in the past so that it
remains stable. Since our links are fairly short-lived, we aren't doing this
now so the benefits of this caching are marginal at best.